tools/knowledge-pipeline/pyproject.toml
[project]
name = "knowledge-pipeline"
version = "0.1.0"
description = "Personal knowledge ingestion pipeline"
requires-python = ">=3.11"
dependencies = [
    "click>=8.1",
    "pydantic>=2.0",
    "instructor>=1.0",
    "structlog>=24.0",
    "httpx>=0.27",
    "keyring>=25.0",
    "faster-whisper>=1.0",
    "rapidfuzz>=3.0",
    "sqlite-vec>=0.1",
    "PyYAML>=6.0",
]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0",
    "pytest-bdd>=7.0",
    "ruff>=0.4",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/kp"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
